server/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
nltk
python-dotenv